fix pickingId's type when max picking > 65536

+++ packages/vector-packer/src/pack/util/array.js
@@ -40,5 +40,5 @@ export function getPosArrayType(max) {
 export function getUnsignedArrayType(max) {
     if (max < 256) return Uint8Array;
     if (max < 65536) return Uint16Array;
-    return Uint32Array;
+    return Float32Array;
 }
